# apps/BE/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Todo API",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
      "Todo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "done": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["id", "title", "done", "createdAt", "updatedAt"]
      },
      "CreateTodo": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": ["title"]
      },
      "SetTodoDone": {
        "type": "object",
        "properties": {
          "done": {
            "type": "boolean"
          }
        },
        "required": ["done"]
      },
      "TodoIdParams": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["id"]
      }
    },
    "parameters": {}
  },
  "paths": {
    "/todos": {
      "get": {
        "responses": {
          "200": {
            "description": "List of todos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "requestBody": {
          "description": "Todo to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodo"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created todo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          }
        }
      }
    },
    "/todos/{id}/done": {
      "patch": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "description": "Done state to set",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetTodoDone"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated todo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "404": {
            "description": "Todo not found"
          }
        }
      }
    }
  }
}