# apps/FE/src/lib/api/generated/model/todo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Todo API
 * OpenAPI spec version: 0.1.0
 */

export interface Todo {
  id: string
  title: string
  done: boolean
  createdAt: string
  updatedAt: string
}